Refactoring to simplify suite setup
