--- junitperf-junit5-suite/src/main/java/com/github/noconnor/junitperf/suite/JUnitPerfSuiteSetup.java
@@ -1,12 +1,0 @@
-package com.github.noconnor.junitperf.suite;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface JUnitPerfSuiteSetup {
-
-}
